+++ saleor/graphql/menu/tests/test_menu.py
@@ -426,21 +426,24 @@ def test_query_menu_items_with_sort(
         assert menu_items[order]["node"]["name"] == menu_item_name
 
 
-def test_menu_item_query_static_url(user_api_client, menu_item):
-    query = """
-    query menuitem($id: ID!) {
-        menuItem(id: $id) {
-            name
-            url
-            category {
-                id
-            }
-            page {
-                id
-            }
+QUERY_MENU_ITEM = """
+query menuitem($id: ID!) {
+    menuItem(id: $id) {
+        name
+        url
+        category {
+            id
+        }
+        page {
+            id
         }
     }
-    """
+}
+"""
+
+
+def test_menu_item_query_static_url(user_api_client, menu_item):
+    query = QUERY_MENU_ITEM
     menu_item.url = "http://example.com"
     menu_item.save()
     variables = {"id": graphene.Node.to_global_id("MenuItem", menu_item.pk)}
@@ -451,6 +454,55 @@ def test_menu_item_query_static_url(user_api_client, menu_item):
     assert data["url"] == menu_item.url
     assert not data["category"]
     assert not data["page"]
+
+
+def test_menu_item_query_staff_with_permission_gets_all_pages(
+    staff_api_client, permission_manage_pages, menu_item, page
+):
+    # given
+    staff_api_client.user.user_permissions.add(permission_manage_pages)
+    variables = {"id": graphene.Node.to_global_id("MenuItem", menu_item.pk)}
+
+    page.is_published = False
+    page.save(update_fields=["is_published"])
+
+    menu_item.page = page
+    menu_item.save(update_fields=["page"])
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_MENU_ITEM, variables)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["menuItem"]
+
+    assert data["name"] == menu_item.name
+    assert data["url"] == menu_item.url
+    assert data["page"]["id"] == graphene.Node.to_global_id("Page", page.id)
+
+
+def test_menu_item_query_staff_without_permission_gets_only_published_pages(
+    staff_api_client, permission_manage_pages, menu_item, page
+):
+    # given
+    variables = {"id": graphene.Node.to_global_id("MenuItem", menu_item.pk)}
+
+    page.is_published = False
+    page.save(update_fields=["is_published"])
+
+    menu_item.page = page
+    menu_item.save(update_fields=["page"])
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_MENU_ITEM, variables)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["menuItem"]
+
+    assert data["name"] == menu_item.name
+    assert data["url"] == menu_item.url
+    assert data["page"] is None
 
 
 def test_create_menu(