+++ saleor/graphql/product/tests/test_product_sorting_attributes.py
@@ -4,8 +4,9 @@ from decimal import Decimal
 import graphene
 import pytest
 
+from ....attribute import AttributeType, models as attribute_models
+from ....attribute.utils import associate_attribute_values_to_instance
 from ....product import models as product_models
-from ....product.utils.attributes import associate_attribute_values_to_instance
 from ...tests.utils import get_graphql_content
 
 HERE = os.path.realpath(os.path.dirname(__file__))
@@ -51,7 +52,7 @@ def products_structures(category, channel_USD):
 
     assert product_models.Product.objects.count() == 0
 
-    in_multivals = product_models.AttributeInputType.MULTISELECT
+    in_multivals = attribute_models.AttributeInputType.MULTISELECT
 
     pt_apples, pt_oranges, pt_other = list(
         product_models.ProductType.objects.bulk_create(
@@ -70,13 +71,20 @@ def products_structures(category, channel_USD):
     )
 
     colors_attr, trademark_attr, dummy_attr = list(
-        product_models.Attribute.objects.bulk_create(
+        attribute_models.Attribute.objects.bulk_create(
             [
-                product_models.Attribute(
-                    name="Colors", slug="colors", input_type=in_multivals
+                attribute_models.Attribute(
+                    name="Colors",
+                    slug="colors",
+                    input_type=in_multivals,
+                    type=AttributeType.PRODUCT_TYPE,
                 ),
-                product_models.Attribute(name="Trademark", slug="trademark"),
-                product_models.Attribute(name="Dummy", slug="dummy"),
+                attribute_models.Attribute(
+                    name="Trademark", slug="trademark", type=AttributeType.PRODUCT_TYPE
+                ),
+                attribute_models.Attribute(
+                    name="Dummy", slug="dummy", type=AttributeType.PRODUCT_TYPE
+                ),
             ]
         )
     )
@@ -552,8 +560,10 @@ def test_sort_product_not_having_attribute_data(api_client, category, count_quer
     )
 
     # Assign an attribute to the product type
-    attribute = product_models.Attribute.objects.create(name="Kind", slug="kind")
-    value = product_models.AttributeValue.objects.create(
+    attribute = attribute_models.Attribute.objects.create(
+        name="Kind", slug="kind", type=AttributeType.PRODUCT_TYPE
+    )
+    value = attribute_models.AttributeValue.objects.create(
         name="Value", slug="value", attribute=attribute
     )
     product_type.product_attributes.add(attribute)
@@ -621,8 +631,8 @@ def test_sort_product_by_attribute_using_attribute_having_no_products(
     """Ensure passing an empty attribute ID as sorting field does nothing."""
 
     query = QUERY_SORT_PRODUCTS_BY_ATTRIBUTE
-    attribute_without_products = product_models.Attribute.objects.create(
-        name="Colors 2", slug="colors-2"
+    attribute_without_products = attribute_models.Attribute.objects.create(
+        name="Colors 2", slug="colors-2", type=AttributeType.PRODUCT_TYPE
     )
 
     attribute_id: str = graphene.Node.to_global_id(