+++ saleor/page/error_codes.py
@@ -7,3 +7,5 @@ class PageErrorCode(Enum):
     NOT_FOUND = "not_found"
     REQUIRED = "required"
     UNIQUE = "unique"
+    DUPLICATED_INPUT_ITEM = "duplicated_input_item"
+    ATTRIBUTE_ALREADY_ASSIGNED = "attribute_already_assigned"