+++ saleor/page/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 from ..core.db.fields import SanitizedJSONField
 from ..core.models import ModelWithMetadata, PublishableModel, PublishedQuerySet
-from ..core.permissions import PagePermissions
+from ..core.permissions import PagePermissions, PageTypePermissions
 from ..core.sanitizers.editorjs_sanitizer import clean_editor_js
 from ..core.utils.translations import TranslationProxy
 from ..seo.models import SeoModel, SeoModelTranslation
@@ -17,6 +17,9 @@ class PagePublishedQuerySet(PublishedQuerySet):
 class Page(ModelWithMetadata, SeoModel, PublishableModel):
     slug = models.SlugField(unique=True, max_length=255)
     title = models.CharField(max_length=250)
+    page_type = models.ForeignKey(
+        "PageType", related_name="pages", on_delete=models.CASCADE
+    )
     content = models.TextField(blank=True)
     content_json = SanitizedJSONField(
         blank=True, default=dict, sanitizer=clean_editor_js
@@ -60,3 +63,17 @@ class PageTranslation(SeoModelTranslation):
 
     def __str__(self):
         return self.title
+
+
+class PageType(ModelWithMetadata):
+    name = models.CharField(max_length=250)
+    slug = models.SlugField(max_length=255, unique=True, allow_unicode=True)
+
+    class Meta:
+        ordering = ("slug",)
+        permissions = (
+            (
+                PageTypePermissions.MANAGE_PAGE_TYPES_AND_ATTRIBUTES.codename,
+                "Manage page types and attributes.",
+            ),
+        )